requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#

requests==2.31.0

pyyaml==6.0

websocket-client==1.5.3
pytz~=2023.3
urllib3~=2.0.3
tqdm~=4.65.0